+++ BookClubWeb/IBooksRepository.cs
@@ -5,7 +5,7 @@ namespace BookClubWeb
     public interface IBooksRepository
     {
         public IEnumerable<Books> GetAllBooks();
-        public Books GetBook(int id);
+        public Books GetBook(double id);
         public void UpdateBook(Books books);
         public void InsertBook(Books bookstoInsert);
         public IEnumerable<Category> GetCategories();
